rewrite EmpiricalPotential
Stored EmpiricalPotential nodes should not rely on functions from modules which may not be there upon execution of the calculation. Therefore, the node should create and store all required data on initialisation.
Also, file contents should be stored in the repository, not as database attributes.
Finally the potential creation modules are now accessed via the entry point system, rather than requiring them to be in a specific location.