fix a bug in trajectory parsing, and allow for charge to be parsed
in `lammps.force` and `lammps.optimize`, the final trajectory was not be parsed, because of a bug in `read_lammps_positions_and_forces_txt`. Alos the wrong `dump` command was being set for `lammps.force`.
Also, if the potential denotes `atom_style='charge'`, then `q` will be appended to the `dump` keywords.