Bump the aiida-core version to 2.x
Removing the reentry command from the tests as it is not needed for aiida-2.x

Removing the md_multi capability of reading the restart for each stage, as it is incompatible with aiida-2.x

Fixing the last of the tests that was blocking the aiida-2.x adoption.

Removing unnecessary tests and adding an older version of lammps for testing.

Fixing an issue in which the lammps version could not be parsed properly when it was not only a datetime.