`BaseLammpsCalculation`: Do not set `CodeInfo.withmpi`
The plugin was setting `CodeInfo.withmpi` based on the inputs from
`metadata.options.withmpi`. This is not necessary as the metadata option
is already handled by the `CalcJob` base class. The plugin implementation
should only set `CodeInfo.withmpi` if it wants to _force_ the code to be
run with MPI enabled or disabled.

The `CalcJob` base implementation used to not distinguish between the
two, but in https://github.com/aiidateam/aiida-core/pull/5922 the
handling of MPI was improved. A `Code` can now define whether it
should be run with MPI or not and so the `CalcJob` class now has to
raise if the MPI setting of the code and the plugin clash. The `withmpi`
option, if explicitly set by the user, is also checked.

Since the plugin doesn't intend to enforce MPI or not, the setting on
the `CodeInfo` is removed.