`LAMMPSBaseParser`: Fix parsing for nodes with the `script` input
The `BaseLammpsCalculation` plugin was recently updated to allow
specifying the exact script to run through the `script` input node. If
the default parser is used in this case, which is the `LAMMPSBaseParser`,
a non-zero exit code would always be returned because some of the output
files that the parser wouldn't be there. These output files are normally
written by the script that the plugin itself would create, but if a
complete script is taken from inputs, it cannot be guaranteed that the
script will produce these same outputs.

The parser is updated to only return an exit code if an expected output
file is not in the retrieved files if the `script` input was not
defined. In addition, the `total_wall_time` key, if parsed by the log
file parser, is parsed and converted to seconds and added as the
`total_wall_time_seconds` key.