`LAMMPSBaseParser`: Fix the parsing of the performance data
The parser attempts to extract the measure of the steps per second from
the logfile through a regex to match the following syntax:

    Performance: 39.271 ns/day, 0.611 hours/ns, 45452.422 timesteps/s

However, this line can also appear in the form:

    Performance: 140354.357 tau/day, 324.894 timesteps/s, 162.447 Matom-step/s

The regex is updated to be indifferent as to the position of the measure
and simply look for the float preceding the `timesteps/s` marker. A test
is added with an example log file with the latter format.

Due to the fixed parsing the reference data of the `test_parser_log`
test was out of date. The test used a manual manner of loading the
reference file and comparing it. This is replaced with the
`data_regression` fixture, which does this automatically and can be used
to automatically update the reference file.