`LammpsBaseCalculation`: Add the `files` and `filenames` inputs
The `files` input namespace allows a user to pass in additional files to
be written to the working directory. By default, the filename to write
the content of each `SinglefileData` node to, defaults to the filename
attribute of the node itself. A validator ensures that the list of
filenames is unique, because otherwise some files would end up
overwriting others.

If this is the case, the optional `filenames` input can be specified to
override the default filename in order to avoid these clashes.