the parser works when using many thermo_style
before the parser was adding all the data in the same array, and parsing
the numbers with the last headers found messing up if many headers and
data columns were present in the output file.
Now, if many header are present, the old data is forgotten as soon as a
new Step ... header is found, and only the last one will be parsed in
the output. This allows to parse output from LammpsRawCalculation
with different thermo_style without errors, as the one found in
https://docs.lammps.org/compute_heat_flux.html#example-input-file

added test